server-smart-caregiver/pyproject.toml
[project]
name = "fastapi-worker"
version = "0.1.0"
description = "Python fastapi example"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "markupsafe",
    "jinja2",
    "sqlalchemy[asyncio]>=2.0.49",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.1",
    "authlib>=1.7.0",
]

[dependency-groups]
dev = [
    "workers-py",
    "workers-runtime-sdk",
    "uvicorn",
    "python-dotenv"
]
